feat: add app.py
